+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Experiment module: Duplicate proposals. [#659](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/659)
+- E-mail notification setting for experiments. [#667](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/667)
+- E-mail notification setting for proposals. [#668](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/668)
 
 ## [2.22.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/43)